css fixes and small changes
